Update package_update auth to support collaborators
This is a critical auth function as it ultimately affects authorization
on many other actions at the dataset, resource, resource view and
DataStore level.

In essence we call the core package_update authorization and if it
fails we perform an additional check against the collaborators
(collaborators permissions are always in addition to org-based ones, ie
is not possible to pass core authorization and then have it removed by
the collaborators one).

Due to ckan/ckan#4828 it was not possible to use the chained_auth_function
decorator.

Added extensive tests.